+++ sdk/schemaregistry/azure-schemaregistry/azure/schemaregistry/_common/_constants.py
@@ -34,4 +34,13 @@ class SchemaFormat(str, Enum):
     AVRO = "avro"
     """Represents the Apache Avro schema format."""
 
-SUPPORTED_API_VERSIONS = ["2021-10"]
+class ApiVersion(str, Enum):
+    """
+    Represents the Schema Registry API version to use for requests.
+    """
+
+    V2020_10 = "2020-10"
+    """This is the default version."""
+
+
+DEFAULT_VERSION = ApiVersion.V2020_10