+++ sdk/schemaregistry/azure-schemaregistry/azure/schemaregistry/_schema_registry_client.py
@@ -26,7 +26,7 @@
 from typing import Any, TYPE_CHECKING, Union
 
 from ._utils import get_http_request_kwargs
-from ._common._constants import SchemaFormat, SUPPORTED_API_VERSIONS
+from ._common._constants import SchemaFormat, DEFAULT_VERSION
 from ._common._schema import Schema, SchemaProperties
 from ._common._response_handlers import (
     _parse_response_schema,
@@ -48,9 +48,8 @@ class SchemaRegistryClient(object):
      For example: my-namespace.servicebus.windows.net.
     :param credential: To authenticate managing the entities of the SchemaRegistry namespace.
     :type credential: ~azure.core.credentials.TokenCredential
-    :keyword api_version: The Schema Registry service API version to use for requests.
+    :keyword str api_version: The Schema Registry service API version to use for requests.
         Default value and only accepted value currently is "2021-10".
-    :paramtype api_version: str
 
     .. admonition:: Example:
 
@@ -65,16 +64,12 @@ class SchemaRegistryClient(object):
 
     def __init__(self, fully_qualified_namespace, credential, **kwargs):
         # type: (str, TokenCredential, Any) -> None
-        api_version = kwargs.get("api_version")
-        if api_version and api_version not in SUPPORTED_API_VERSIONS:
-            versions = "\n".join(SUPPORTED_API_VERSIONS)
-            raise ValueError(
-                "Unsupported API version '{}'. Please select from:\n{}".format(
-                    api_version, versions
-                )
-            )
+        api_version = kwargs.pop("api_version", DEFAULT_VERSION)
         self._generated_client = AzureSchemaRegistry(
-            credential=credential, endpoint=fully_qualified_namespace, **kwargs
+            credential=credential,
+            endpoint=fully_qualified_namespace,
+            api_version=api_version,
+            **kwargs
         )
 
     def __enter__(self):
@@ -94,7 +89,12 @@ class SchemaRegistryClient(object):
         self._generated_client.close()
 
     def register_schema(
-        self, group_name, schema_name, schema_definition, format, **kwargs  # pylint:disable=redefined-builtin
+        self,
+        group_name,
+        schema_name,
+        schema_definition,
+        format,
+        **kwargs  # pylint:disable=redefined-builtin
     ):
         # type: (str, str, str, Union[str, SchemaFormat], Any) -> SchemaProperties
         """
@@ -161,13 +161,20 @@ class SchemaRegistryClient(object):
 
         """
         http_request_kwargs = get_http_request_kwargs(kwargs)
-        request = schema_rest.build_get_by_id_request(schema_id=schema_id, **http_request_kwargs)
+        request = schema_rest.build_get_by_id_request(
+            schema_id=schema_id, **http_request_kwargs
+        )
         response = self._generated_client.send_request(request, **kwargs)
         response.raise_for_status()
         return _parse_response_schema(response)
 
     def get_schema_properties(
-        self, group_name, schema_name, schema_definition, format, **kwargs  # pylint:disable=redefined-builtin
+        self,
+        group_name,
+        schema_name,
+        schema_definition,
+        format,
+        **kwargs  # pylint:disable=redefined-builtin
     ):
         # type: (str, str, str, Union[str, SchemaFormat], Any) -> SchemaProperties
         """